# packages/contracts/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/characters/{character_id}/settlements/latest": {
      "get": {
        "operationId": "SettlementController_getLatest",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "最近一次离线结算摘要。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeSettlementSummary"
                }
              }
            }
          },
          "404": {
            "description": "资源未找到。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "读取角色最近一次离线结算摘要",
        "tags": [
          "settlement"
        ]
      }
    },
    "/api/v1/characters/{character_id}/settlements/{settlement_id}": {
      "get": {
        "operationId": "SettlementController_getById",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "settlement_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "指定离线结算摘要。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeSettlementSummary"
                }
              }
            }
          },
          "404": {
            "description": "资源未找到。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "读取指定离线结算摘要",
        "tags": [
          "settlement"
        ]
      }
    }
  },
  "info": {
    "title": "洞天 API",
    "description": "洞天服务端 REST API。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "required": [
          "request_id",
          "server_time"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "server_time": {
            "type": "string",
            "format": "date-time"
          },
          "config_version": {
            "type": "string"
          },
          "state_version": {
            "type": "integer"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error",
          "meta"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message_key",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message_key": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "SettlementRewardItem": {
        "type": "object",
        "required": [
          "item_id",
          "quantity"
        ],
        "properties": {
          "item_id": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          }
        }
      },
      "SettlementRewards": {
        "type": "object",
        "required": [
          "cultivation_xp",
          "skill_xp",
          "items"
        ],
        "properties": {
          "cultivation_xp": {
            "type": "string"
          },
          "skill_xp": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettlementRewardItem"
            }
          }
        }
      },
      "SettlementTimelineEntry": {
        "type": "object",
        "required": [
          "segment_index",
          "queue_entry_id",
          "action_config_id",
          "from_at",
          "to_at",
          "completed_cycles",
          "inputs",
          "outputs",
          "xp_changes",
          "transition_reason",
          "snapshot"
        ],
        "properties": {
          "segment_index": {
            "type": "integer"
          },
          "queue_entry_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_config_id": {
            "type": "string"
          },
          "from_at": {
            "type": "string",
            "format": "date-time"
          },
          "to_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_cycles": {
            "type": "string"
          },
          "inputs": {
            "type": "object",
            "additionalProperties": true
          },
          "outputs": {
            "type": "object",
            "additionalProperties": true
          },
          "xp_changes": {
            "type": "object",
            "additionalProperties": true
          },
          "transition_reason": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "SettlementLedgerEntry": {
        "type": "object",
        "required": [
          "entry_id",
          "transaction_id",
          "asset_type",
          "asset_id",
          "delta",
          "balance_after",
          "reason_code",
          "reference_type",
          "reference_id",
          "config_version",
          "created_at"
        ],
        "properties": {
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid"
          },
          "asset_type": {
            "type": "string"
          },
          "asset_id": {
            "type": "string"
          },
          "delta": {
            "type": "string"
          },
          "balance_after": {
            "type": "string"
          },
          "reason_code": {
            "type": "string"
          },
          "reference_type": {
            "type": "string"
          },
          "reference_id": {
            "type": "string"
          },
          "config_version": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LatestSettlementSummary": {
        "type": "object",
        "required": [
          "settlement_id",
          "character_id",
          "as_of",
          "from_at",
          "requested_until",
          "effective_until",
          "effective_time_us",
          "capped_time_us",
          "continuation_required",
          "status",
          "summary",
          "rewards",
          "timeline",
          "ledger_entries"
        ],
        "properties": {
          "settlement_id": {
            "type": "string",
            "format": "uuid"
          },
          "character_id": {
            "type": "string",
            "format": "uuid"
          },
          "as_of": {
            "type": "string",
            "format": "date-time"
          },
          "from_at": {
            "type": "string",
            "format": "date-time"
          },
          "requested_until": {
            "type": "string",
            "format": "date-time"
          },
          "effective_until": {
            "type": "string",
            "format": "date-time"
          },
          "effective_time_us": {
            "type": "string"
          },
          "capped_time_us": {
            "type": "string"
          },
          "continuation_required": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "object",
            "additionalProperties": true
          },
          "rewards": {
            "$ref": "#/components/schemas/SettlementRewards"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettlementTimelineEntry"
            }
          },
          "ledger_entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettlementLedgerEntry"
            }
          }
        }
      },
      "LatestSettlementResponse": {
        "type": "object",
        "required": [
          "settlement"
        ],
        "properties": {
          "settlement": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LatestSettlementSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "SuccessEnvelopeSettlementSummary": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LatestSettlementResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      }
    }
  }
}